poll-starter/requirements.txt
psycopg2-binary==2.8.5
python-dotenv==0.15.0
